requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
Django>=4.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
inflection==0.5.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0

psycopg[binary]>=3.2
gunicorn>=21.0
django-cors-headers==4.3.0